CLOUD-59100 Rest and identity client moved to cloud common module
