--- autoscale-api/src/main/java/com/sequenceiq/periscope/api/exception/SSLConnectionException.java
@@ -1,14 +1,0 @@
-package com.sequenceiq.periscope.api.exception;
-
-public class SSLConnectionException extends RuntimeException {
-
-    public SSLConnectionException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public SSLConnectionException(String message) {
-        super(message);
-    }
-}
-
-
