+++ cloud-common/src/main/java/com/sequenceiq/cloudbreak/client/AccessToken.java
