+++ cloud-common/src/main/java/com/sequenceiq/cloudbreak/client/ConfigKey.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.client.config;
+package com.sequenceiq.cloudbreak.client;
 
 
 public class ConfigKey {
