+++ cloud-common/src/main/java/com/sequenceiq/cloudbreak/client/KeyStoreUtil.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.util;
+package com.sequenceiq.cloudbreak.client;
 
 import java.io.BufferedReader;
 import java.io.File;
