+++ cloud-common/src/main/java/com/sequenceiq/cloudbreak/client/RestClientUtil.java
@@ -1,11 +1,14 @@
 package com.sequenceiq.cloudbreak.client;
 
+import static org.terracotta.modules.ehcache.store.TerracottaClusteredInstanceFactory.LOGGER;
+
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 import javax.net.ssl.SSLContext;
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.ClientBuilder;
+import javax.ws.rs.client.WebTarget;
 
 import org.apache.http.config.Registry;
 import org.apache.http.config.RegistryBuilder;
@@ -13,21 +16,19 @@ import org.apache.http.conn.socket.ConnectionSocketFactory;
 import org.apache.http.conn.socket.PlainConnectionSocketFactory;
 import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
 import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
+import org.apache.http.ssl.SSLContexts;
 import org.glassfish.jersey.apache.connector.ApacheClientProperties;
 import org.glassfish.jersey.apache.connector.ApacheConnectorProvider;
 import org.glassfish.jersey.client.ClientConfig;
 import org.glassfish.jersey.client.ClientProperties;
 import org.glassfish.jersey.client.JerseyClientBuilder;
+import org.glassfish.jersey.client.authentication.HttpAuthenticationFeature;
 import org.glassfish.jersey.filter.LoggingFilter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import com.sequenceiq.cloudbreak.client.config.ConfigKey;
-import com.sequenceiq.cloudbreak.client.security.CertificateTrustManager;
+public class RestClientUtil {
 
-public class RestClient {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(RestClient.class);
+    private static final String USER = "cbadmin";
+    private static final String PASSWORD = "cbadmin";
 
     // apache http connection pool defaults are constraining
     // https://hc.apache.org/httpcomponents-client-ga/tutorial/html/connmgmt.html
@@ -36,21 +37,69 @@ public class RestClient {
 
     private static ConcurrentMap<ConfigKey, Client> clients = new ConcurrentHashMap<>();
 
-    private RestClient() {
+    private RestClientUtil() {
+    }
+
+    public static synchronized Client get() {
+        return get(new ConfigKey(false, false));
     }
 
     public static synchronized Client get(ConfigKey configKey) {
         Client client = clients.get(configKey);
         if (client == null) {
-            client = constructClient(configKey);
+            client = createClient(configKey);
             clients.put(configKey, client);
         }
         LOGGER.info("RestClient cache size: {}, key: {}, fetched client: {}", clients.size(), configKey, client);
         return client;
-
     }
 
-    private static Client constructClient(ConfigKey configKey) {
+    public static Client createClient(String serverCert, String clientCert, String clientKey) throws Exception {
+        return createClient(serverCert, clientCert, clientKey, false, null);
+    }
+
+    public static Client createClient(String serverCert, String clientCert, String clientKey, boolean debug, Class debugClass) throws Exception {
+        SSLContext sslContext = SSLContexts.custom()
+                .loadTrustMaterial(KeyStoreUtil.createTrustStore(serverCert), null)
+                .loadKeyMaterial(KeyStoreUtil.createKeyStore(clientCert, clientKey), "consul".toCharArray())
+                .build();
+
+        LOGGER.info("Constructing jax rs client for config: server cert: {}, client cert: {}, debug: {}", serverCert, clientCert, debug);
+        ClientConfig config = new ClientConfig();
+        config.property(ClientProperties.FOLLOW_REDIRECTS, "false");
+
+        RegistryBuilder<ConnectionSocketFactory> registryBuilder = RegistryBuilder.create();
+        registryBuilder.register("http", PlainConnectionSocketFactory.getSocketFactory());
+        registryBuilder.register("https", new SSLConnectionSocketFactory(sslContext));
+
+        PoolingHttpClientConnectionManager connectionManager =
+                new PoolingHttpClientConnectionManager(registryBuilder.build());
+
+        connectionManager.setMaxTotal(MAX_TOTAL_CONNECTION);
+        connectionManager.setDefaultMaxPerRoute(MAX_PER_ROUTE_CONNECTION);
+
+        // tell the jersey config about the connection manager
+        config.property(ApacheClientProperties.CONNECTION_MANAGER, connectionManager);
+        config.connectorProvider(new ApacheConnectorProvider());
+
+        ClientBuilder builder = JerseyClientBuilder.newBuilder().withConfig(config);
+
+        if (debug) {
+            builder = builder.register(new LoggingFilter(java.util.logging.Logger.getLogger(debugClass.getName()),
+                    true));
+        }
+
+        Client client = builder.build();
+        LOGGER.info("Jax rs client has been constructed: {}, sslContext: {}", client, sslContext);
+        return client;
+    }
+
+    public static WebTarget createTarget(Client client, String address) {
+        HttpAuthenticationFeature feature = HttpAuthenticationFeature.basic(USER, PASSWORD);
+        return client.target(address).register(feature);
+    }
+
+    private static Client createClient(ConfigKey configKey) {
         LOGGER.info("Constructing jax rs client: {}", configKey);
         ClientConfig config = new ClientConfig();
         config.property(ClientProperties.FOLLOW_REDIRECTS, "false");
@@ -78,7 +127,7 @@ public class RestClient {
         ClientBuilder builder = JerseyClientBuilder.newBuilder().withConfig(config);
 
         if (configKey.isDebug()) {
-            builder = builder.register(new LoggingFilter(java.util.logging.Logger.getLogger(RestClient.class.getName()), true));
+            builder = builder.register(new LoggingFilter(java.util.logging.Logger.getLogger(RestClientUtil.class.getName()), true));
         }
 
         Client client = builder.build();
@@ -87,5 +136,4 @@ public class RestClient {
         LOGGER.warn("RestClient has been constructed: {}, client: {}, sslContext: {}", configKey, client, sslContext);
         return client;
     }
-
 }
