+++ cloud-common/src/main/java/com/sequenceiq/cloudbreak/client/SSLConnectionException.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.api.exception;
+package com.sequenceiq.cloudbreak.client;
 
 public class SSLConnectionException extends RuntimeException {
