+++ core-api/build.gradle
@@ -30,6 +30,7 @@ dependencies {
   testCompile group: 'org.springframework.boot',  name: 'spring-boot-starter-test',       version: springBootVersion
   testCompile group: 'junit',                     name: 'junit',                          version: junitVersion
 
+  compile project(':cloud-common')
 }
 
 task testJar(type: Jar, dependsOn: testClasses) {
