+++ core/build.gradle
@@ -48,7 +48,6 @@ dependencies {
     compile group: 'org.apache.commons',            name: 'commons-lang3',                  version: apacheCommonsLangVersion
     compile group: 'org.freemarker',                name: 'freemarker',                     version: freemarkerVersion
     compile group: 'org.postgresql',                name: 'postgresql',                     version: '9.3-1101-jdbc41'
-    compile group: 'com.sequenceiq',                name: 'consul-api',                     version: '1.10'
     compile group: 'org.codehaus.jettison',         name: 'jettison',                       version: '1.3.5'
     compile group: 'com.thoughtworks.xstream',      name: 'xstream',                        version: '1.4.7'
     compile group: 'com.github.fommil',             name: 'openssh',                        version: '1.0'
